Fix sessionStorage not being cleaned on refresh

+++ src/wrappers/SessionCleaner.tsx
@@ -2,9 +2,16 @@ import { PropsWithChildren, useEffect } from "react";
 
 export default function SessionCleaner({ children }: PropsWithChildren) {
   useEffect(() => {
-    return () => {
-      console.log("<SessionCleaner /> unmounted!");
-      sessionStorage.clear();
+    window.onbeforeunload = function () {
+      sessionStorage.setItem("refresh-initiator", window.location.href);
+    };
+    window.onload = function () {
+      if (
+        window.location.href === sessionStorage.getItem("refresh-initiator")
+      ) {
+        console.log("Cleaning sessionStorage..");
+        sessionStorage.clear();
+      }
     };
   }, []);
   return <>{children}</>;
